FT2-43: Made few changes and added files to implement OOPs.
