+++ advanced2/LoadEnv.php
@@ -1,0 +1,11 @@
+<?php
+require 'vendor/autoload.php';
+use Dotenv\Dotenv;
+
+class LoadEnv {
+  public static function loadDotenv(){
+    $dotenv = Dotenv::createImmutable(__DIR__);
+    $dotenv->load();
+  } 
+}
+?>
