+++ advanced2/ValidateMail.php
@@ -1,16 +1,11 @@
 <?php
 
 use GuzzleHttp\Client;
-use Dotenv\Dotenv;
-require 'vendor/autoload.php';
-require 'send.php';
-$dotenv = Dotenv::createImmutable(__DIR__);
-$dotenv->load();
-
+require_once 'Mailer.php';
 /**
  * Class to validate Mail and store related info.
  */
-class ValidateMail {
+class ValidateMail extends Mailer{
   public $email;
   public $emailErrorMessage;
   public $validError;
@@ -21,15 +16,14 @@ class ValidateMail {
   }
 
   function validate() {
-    // Regular Expression to check if the email is in a valid format or not. 
-    $regularExp = '/^[\w._]+@[\w]+(\.[a-z]{2,}){0,2}$/i';
-
-    if (!preg_match($regularExp, $this->email)) {
+    // Check if the email is in valid format or not.
+    if (!(filter_var($this->email, FILTER_VALIDATE_EMAIL))) {
       $this->emailError = 1;
       $this->emailErrorMessage = "Invalid email address format!";
     }
   
     if ($this->emailError != 1) {
+      LoadEnv::loadDotenv();
       $client = new Client();
       $access_key = $_ENV['ACCESS_KEY'];
       $response = $client->request('GET', 'https://emailvalidation.abstractapi.com/v1/?api_key=' . $access_key . '&email=' . $this->email);
@@ -46,21 +40,10 @@ class ValidateMail {
       }
       else {
         // If everything is okay then call the send.php file to send the mail.
-        sendMail($this->email);
+        $this->sendMail($this->email);
         header('Location: message.php');
         exit;
       }
     }
-  }
-}
-$validError = $emailErrorMessage = "";
-if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-  $email = $_POST['email'];
-  $evalObj = new ValidateMail($email);
-  $evalObj->validate();
-  if($evalObj->emailError){
-    $validError = $evalObj->validError;
-    $emailErrorMessage = $evalObj->emailErrorMessage;
-    require 'error.php';
   }
 }
